Update Connect to a MySQL database.php
Here is some sample PHP code to connect to a MySQL database. You will need to replace the values for $servername, $username, $password, and $dbname with the appropriate values for your MySQL server. You should also check for errors when sending queries to the database, and close the connection when you are done using it.