Create Arithmetics computations.php
In PHP, you can calculate the value of two variables by using mathematical operators such as + (addition), - (subtraction), * (multiplication), / (division), and % (modulus). Here is an example of how to calculate two variables in PHP:

(Code above)

In this example, the variables $x and $y are assigned 5 and 2, respectively. The $result variable is then set the value of $x + $y, which is 7. The echo statement is used to print the result to the screen. You can also use other mathematical operators to perform other calculations like the above code.  This will give you the result of subtraction, multiplication, division, and modulus respectively.