Create upload an image to a folder on a server.php
Here is an example of PHP code that can be used to upload an image to a folder on a server. This code assumes that the form used to submit the image has a file input field with the name "image" and that there is a folder called "images" in the same directory as the PHP file. It also check for file type and size of image before uploading to folder.