Create CRUD.php
CRUD stands for Create, Read, Update, and Delete. It is a set of operations commonly used in database applications. Here is an example of PHP code that performs CRUD operations on a MySQL database: 

Please note that the above code is just a simple example, it is important to consider security issues, such as prepared statements, when handling user inputs.