Create MYSLi CRUD (PDO).php
CRUD stands for Create, Read, Update, and Delete, which are the four basic functions of persistent storage. Here is an example of how to perform CRUD operations using PHP's PDO (PHP Data Objects) library: Please note that this is just an example and it should be adjusted to the specific requirements and the database.