Create Here is an example of a PHP script that inserts a new contact into a MySQLi database, with some basic input validation.php
Here is an example of a PHP script that inserts a new contact into a MySQLi database, with some basic input validation:
This script first creates a connection to the MySQLi database using the provided connection details. It then checks if the request method is a post, it then escapes the input values, checks if the fields are empty, checks if the email address is well-formed, and if it is then it inserts the data into the database table 'contacts'. It's important to note that the script should be executed on a secure server and should be protected against SQL injection attacks.