Create PHP and MySQLi to display data in a bar graph.php
Here is an example of using PHP and MySQLi to display data in a bar graph:

This example uses a MySQLi query to retrieve data from a table and store it in an array. Then it iterates through the array, using the "value" field to determine the height of each bar in the graph and a fixed width of 20px and background color is green.

This is a simple example and it's recommended to use a more robust graphing library like Chart.js, Highcharts or D3.js for more advanced functionality and better visual results.