Create PHP and MySQLi to display data in a pie chart.php
Here is an example of how to use PHP and MySQLi to display data in a pie chart:
This code connects to a database, runs a query to get data from a table, formats the data into an array, and then uses the C3 library to create a pie chart from the data. The query retrieves the category and total amount of expenses, grouped by category.