Create php code to read contacts from google spreadsheet.php
Here is an example of how to read data from a Google Spreadsheet using the Google Sheets API and PHP:

Make sure you have enabled the Google Sheets API and have a valid service account credentials.json.
You have to pass the spreadsheet id and range of cells to the get method.

This will print the data in the spreadsheet. You can then iterate over the data and process it as needed.