Create read contacts from Firebase using PHP.php
Here is an example of how to read contacts from Firebase using PHP:

This code uses the Kreait/Firebase-PHP library, which you should have installed via composer. It also uses the Firebase service account credentials in the json file.
You will need to replace google-service-account.json with the path to your Firebase service account credentials.
This code will retrieve the data stored under the "contacts" path in your Firebase Realtime Database and print it to the screen.
Note that You should check the Firebase Database rules to make sure that you have the right permission to read the contacts data.