Create PHP and JavaScript to display data in real-time from a MySQLi database.php
Here is an example of how you could use PHP and JavaScript to display data in real-time from a MySQLi database:In this example, the PHP script retrieves the latest data from the MySQLi database and outputs it as a JSON object. The JavaScript code uses the Fetch API to call the PHP script and update the web page with the new data. The setInterval() function is used to refresh the data every 5 seconds.

Note that you would need to adjust the PHP and JavaScript code to match the specific structure of your MySQLi database and the data you want to display.