Create Images CRUD phpMysqli.php
CRUD (Create, Read, Update, and Delete) operations are commonly used in database management systems and can be implemented using the PHP MySQLi extension. To create an image, you can use the INSERT statement to insert a new record into the database with the image file as a BLOB (binary large object) type. To read an image, you can use the SELECT statement to retrieve the image from the database and display it on the web page. To update an image, you can use the UPDATE statement to modify the image file in the database. To delete an image, you can use the DELETE statement to remove the image from the database.