Create php code to calculate area of a circle.php
This code defines a function called "calculateArea" that takes in one parameter, the radius of the circle. Inside the function, we use the pi() function to get the value of pi and then use the pow() function to calculate the square of the radius. The result of this calculation is returned by the function. Then we define the radius as 5 and call the function to get the area of the circle, and print the result.